+++ sync-for-commerce/src/codatsynccommerce/models/operations/get_configuration.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import configuration as shared_configuration
+from ...models.shared import configuration as shared_configuration
 from typing import Optional