+++ sync-for-commerce/src/codatsynccommerce/models/operations/get_sync_by_id.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import companysyncstatus as shared_companysyncstatus
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import companysyncstatus as shared_companysyncstatus
 from typing import Optional
 
 
@@ -26,8 +25,6 @@ class GetSyncByIDResponse:
     r"""HTTP response status code for this operation"""
     company_sync_status: Optional[shared_companysyncstatus.CompanySyncStatus] = dataclasses.field(default=None)
     r"""Success"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""