+++ sync-for-commerce/src/codatsynccommerce/models/operations/update_connection.py
@@ -3,8 +3,8 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import connection as shared_connection
-from ..shared import updateconnection as shared_updateconnection
+from ...models.shared import connection as shared_connection
+from ...models.shared import updateconnection as shared_updateconnection
 from typing import Optional