+++ sync-for-commerce/src/codatsynccommerce/models/operations/update_visible_accounts_sync_flow.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import visibleaccounts as shared_visibleaccounts
+from ...models.shared import visibleaccounts as shared_visibleaccounts
 from typing import Optional